# retailer_service/src/apis/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "Sarvm Retailer App",
    "description": "Retailer Application API"
  },
...
  "servers": [
    {
      "url": "http://localhost:1215",
      "description": "Local server"
    },
    {
      "url": "http://auth-nlb-e953c9da3a44bec4.elb.ap-south-1.amazonaws.com/",
      "description": "Dev Server"
    },
    {
      "url": "https://api.sarvm.ai/",
      "description": "PRD server"
    },
    {
      "url": "http://auth-nlb-e953c9da3a44bec4.elb.ap-south-1.amazonaws.com/",
      "description": "STG server"
    },
    {
      "url": "http://localhost:3000/",
      "description": "UAT_VPN"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/rms/apis/v1/catalog/{shopid}": {
      "post": {
        "tags": ["Catalog"],
        "summary": "Add Catalog in the system",
        "description": "Add Catalog in the system",
        "operationId": "addCatalog",
        "parameters": [
          {
            "name": "shopid",
            "in": "path",
            "description": "The shopid that needs to be fetched. Use 2 for testing. ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Add Catalog in the system",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Catalog"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Catalog"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Catalog"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSuccess"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSuccess"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError500"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError500"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "tags": ["Catalog"],
        "summary": "Get Catalog",
        "description": "",
        "operationId": "getCatalog",
        "parameters": [
          {
            "name": "shopid",
            "in": "path",
            "description": "The shopid that needs to be fetched. Use 2 for testing. ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSuccess"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError500"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError500"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Catalog"],
        "summary": "Update Catalog",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateCatalog",
        "parameters": [
          {
            "name": "shopid",
            "in": "path",
            "description": "shopid that need to be updated",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update an existent Catalog in the system",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Catalog"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Catalog"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Catalog"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponseSuccess": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "example": []
          }
        },
        "xml": {
          "name": "##default"
        }
      },
      "ApiResponseError500": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "INTERNAL_SERVER_ERROR"
              },
              "originalError": {
                "type": "string",
                "example": "ERROR"
              },
              "message": {
                "type": "string",
                "example": "Internal server error occurred"
              }
            }
          }
        },
        "xml": {
          "name": "##default"
        }
      },
...
      "Catalog": {
        "type": "object",
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "soldBy": {
                  "type": "string",
                  "example": "PICES"
                },
                "sp": {
                  "type": "integer",
                  "format": "int32",
                  "example": "120"
                },
                "grading": {
                  "type": "integer",
                  "format": "int32",
                  "example": "4"
                },
                "min_qty_order": {
                  "type": "integer",
                  "format": "int32",
                  "example": "12"
                },
                "max_qty_order": {
                  "type": "integer",
                  "format": "int32",
                  "example": "45"
                },
                "regular_price": {
                  "type": "integer",
                  "format": "int32",
                  "example": "45"
                },
                "discount_price": {
                  "type": "integer",
                  "format": "int32",
                  "example": "45"
                },
                "min_price_order": {
                  "type": "integer",
                  "format": "int32",
                  "example": "56"
                },
                "quantity": {
                  "type": "integer",
                  "format": "int32",
                  "example": "12"
                },
                "cost_price": {
                  "type": "integer",
                  "format": "int32",
                  "example": "45"
                },
                "sku": {
                  "type": "string",
                  "example": "test"
                },
                "stock_status": {
                  "type": "boolean"
                },
                "published": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token",
        "in": "header",
        "name": "Authorization"
      }
    }
  }
}